+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.18 / 2023-03-24
+  * The following are now parsed property:
+    `die_on_file_error`, `die_on_path_denied`, `die_on_run_error`, `die_on_path_denied`, and `die_on_other_error`.
+
 ## 2.0.17 / 2023-03-22
   * Added `repo` and `git_ref` parameters, so files can be retrieved from git repositories at a given commit or tag.